Refactor evaluator test cases into separate classes. Remove displayWinner() method from board and instead use board.isEndGameState(PieceColor color) for both GameManager and GameOverEvaluator classes. Change board.isEndGameState(PieceColor color) to only return true if it is an end game state for the current player. Use if-else if-else control instead of trailing ifs in GameOverEvaluator.
