+++ src/model/Board.java
@@ -48,25 +48,6 @@ public class Board {
         }
     }
 
-    public PieceColor determineWinner() {
-        final boolean whiteHasNoPieces = this.getNumberOfWhitePieces() == 0;
-        final boolean whiteHasNoMoves = MoveGenerator.getAllPossibleMoves(this, PieceColor.WHITE)
-                .isEmpty();
-        final boolean blackHasNoPieces = this.getNumberOfBlackPieces() == 0;
-        final boolean blackHasNoMoves = MoveGenerator.getAllPossibleMoves(this, PieceColor.BLACK)
-                .isEmpty();
-
-        final boolean whiteWins = blackHasNoPieces || blackHasNoMoves;
-        final boolean blackWins = whiteHasNoPieces || whiteHasNoMoves;
-
-        if (whiteWins) {
-            return PieceColor.WHITE;
-        } else if (blackWins) {
-            return PieceColor.BLACK;
-        }
-        return null;
-    }
-
     @Override
     public boolean equals(Object obj) {
         if (this == obj) {
@@ -209,8 +190,13 @@ public class Board {
     }
 
     public boolean isEndState(PieceColor color) {
-        final boolean outOfPieces = (this.numberOfBlackPieces == 0)
-                || (this.numberOfWhitePieces == 0);
+
+        boolean outOfPieces = false;
+        if (color == PieceColor.BLACK) {
+            outOfPieces = this.numberOfBlackPieces == 0;
+        } else {
+            outOfPieces = this.numberOfWhitePieces == 0;
+        }
 
         if (outOfPieces) {
             return true;
